+++ tests/integration/preview/understand/service/test_model_build.py
@@ -38,7 +38,9 @@ class ModelBuildTestCase(IntegrationTestCase):
                 "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name"
+                "unique_name": "unique_name",
+                "build_duration": null,
+                "error_code": null
             }
             '''
         ))
@@ -103,11 +105,13 @@ class ModelBuildTestCase(IntegrationTestCase):
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "date_updated": "2015-07-30T20:00:00Z",
                         "url": "https://preview.twilio.com/understand/Services/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/ModelBuilds/UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "status": "enqueued",
+                        "status": "failed",
                         "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "date_created": "2015-07-30T20:00:00Z",
                         "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "unique_name": "unique_name"
+                        "unique_name": "unique_name",
+                        "build_duration": null,
+                        "error_code": 23001
                     }
                 ]
             }
@@ -143,7 +147,9 @@ class ModelBuildTestCase(IntegrationTestCase):
                 "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name"
+                "unique_name": "unique_name",
+                "build_duration": null,
+                "error_code": null
             }
             '''
         ))
@@ -173,11 +179,13 @@ class ModelBuildTestCase(IntegrationTestCase):
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_updated": "2015-07-30T20:00:00Z",
                 "url": "https://preview.twilio.com/understand/Services/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/ModelBuilds/UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "enqueued",
+                "status": "completed",
                 "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name"
+                "unique_name": "unique_name",
+                "build_duration": 100,
+                "error_code": null
             }
             '''
         ))